pyproject.toml
[project]
name = "peak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.15.2",
    "apscheduler>=3.11.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pydoc-markdown>=4.8.2",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]
